Move editor states to State project
