+++ src/SimpleLevelEditor/Ui/ChildWindows/EntityEditorWindow.cs
@@ -5,6 +5,7 @@ using SimpleLevelEditor.Formats.Types;
 using SimpleLevelEditor.Formats.Types.EntityConfig;
 using SimpleLevelEditor.Formats.Types.Level;
 using SimpleLevelEditor.State;
+using SimpleLevelEditor.State.Editor;
 using SimpleLevelEditor.State.Level;
 using SimpleLevelEditor.Utils;
 using System.Diagnostics;
@@ -13,23 +14,16 @@ namespace SimpleLevelEditor.Ui.ChildWindows;
 
 public static class EntityEditorWindow
 {
-	public static Entity DefaultEntity { get; private set; } = Entity.CreateDefault();
-
 	public static void Render()
 	{
 		if (ImGui.BeginChild("Edit Entity", default, ImGuiChildFlags.Border))
 		{
 			ImGui.SeparatorText("Edit Entity");
 
-			RenderEntityInputs(LevelEditorState.SelectedEntity ?? DefaultEntity);
+			RenderEntityInputs(LevelEditorState.SelectedEntity ?? EntityEditorState.DefaultEntity);
 		}
 
 		ImGui.EndChild(); // End Object Editor
-	}
-
-	public static void Reset()
-	{
-		DefaultEntity = Entity.CreateDefault();
 	}
 
 	private static void RenderEntityInputs(Entity entity)
